+++ Android/app/src/main/kotlin/com/onewisebit/scpescape/model/repositories/InModeRepository.kt
@@ -4,9 +4,9 @@ import com.onewisebit.scpescape.model.entities.Mode
 
 interface InModeRepository {
 
-    fun getMode(id: Long): Mode
-    fun getMinPlayers(id: Long): Int
-    fun getMaxPlayers(id: Long): Int
+    fun getMode(id: Int): Mode
+    fun getMinPlayers(id: Int): Int
+    fun getMaxPlayers(id: Int): Int
     fun insertMode(mode: Mode)
     fun getAllModes(): List<Mode>
     suspend fun insertAll(modes: List<Mode>)
